requirements.txt
boto3
langchain
langchain-community 
streamlit 
pillow 
faiss-cpu
nltk
openai
python-dotenv
langchain
tiktoken
langchain-openai
watchdog
sentence-transformers
openai
matplotlib
langchain_community